Move the read config file method to one place and imported for every usage
