+++ pipert/core/component_factory.py
@@ -1,28 +1,10 @@
 import argparse
-from typing import Union
-
-import yaml
 import zerorpc
-from yaml.parser import ParserError
-from yaml.scanner import ScannerError
-
 from pipert.core.class_factory import ClassFactory
 from pipert.core.component import BaseComponent
+from pipert.utils.useful_methods import open_config_file
 
 COMPONENTS_FOLDER_PATH = "pipert/contrib/components"
-
-
-def open_config_file(config_path) -> Union[str, dict]:
-    try:
-        with open(config_path) as config_file:
-            return yaml.load(config_file, Loader=yaml.FullLoader)
-    except FileNotFoundError as error:
-        error_msg = error.args[1] + "' {}'".format(config_path)
-    except IsADirectoryError as error:
-        error_msg = "'{}' is a directory not a file".format(config_path)
-    except (ScannerError, ParserError):
-        error_msg = "Expecting yaml file, can't parse the file '{}'".format(config_path)
-    return error_msg
 
 
 if __name__ == '__main__':
